New Artado Search
I rewrote the entire codebase. It's not a update. It's completely a new search engine. The old codebase was huge and badly-written(my bad, sorry). I tried to write clean code. Its still not perfect but better than the old one.

With the new Artado these things have added:
- A better menu
- Custom Themes
- Extensions
- Profiles
and etc.