Landing page ads are working, privacy policy updated, bugs fixed
We added ads on our landing page. We are also planning to add advertisment on results but there are still time for that. I made the basics for anonymous personalized ads, thats why you might see things about it. But right now Artado does not have that. We will add it with result ads. We are still testing the ads. We will add more things about this matter in the future.